MySql.Data.csproj - removed SocketStream Field.cs - changed how we decide if a field is a blob README & Release notes - updated version # Mysql.Data.Tests.csproj - updated nunit used. BaseTests.cs - changed protocol=tcp to protocol=sockets in connection string. DataTypeTests.cs - added test for checking the output of "SHOW COLUMNS" PreparedStatements.cs - Added check to make sure this test only happens on mysql 5.0 or later.

+++ TestSuite/Source/PreparedStatements.cs
@@ -752,6 +752,8 @@ namespace MySql.Data.MySqlClient.Tests
         [Test]
         public void ClosingCommandsProperly()
         {
+            if (version < new Version(5, 0)) return;
+
             execSQL("DROP TABLE IF EXISTS test");
             execSQL("CREATE TABLE test (id INT, name VARCHAR(50))");
